split types into files
